Moves importers into the same module and folder
